+++ Assets/Scenes/TitleScreen.unity
@@ -580,7 +580,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 317765590}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 5ce015c38ab642144b460b3f7d44a1b0, type: 3}
+  m_Script: {fileID: 11500000, guid: 823bea3ac51af2f4c98fbfef5717d6f9, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
 --- !u!1 &381843661
