+++ Assets/Scripts/Menus/TitleMenuActions.cs
@@ -12,14 +12,14 @@ public class TitleMenuActions : MonoBehaviour {
 
 	private ConstantsManager constantsManager;
 	private UIManager uiManager;
-	private SaveGame saveManager;
+	private SaveLoadGame saveManager;
 
 	void Awake() {
 		GameObject gameController = GameObject.FindGameObjectWithTag ("GameController");
 		constantsManager = gameController.GetComponent<ConstantsManager> ();
 		GameObject levelController = GameObject.FindGameObjectWithTag ("LevelController");
 		uiManager = levelController.GetComponent<UIManager> ();
-		saveManager = gameController.GetComponent<SaveGame> ();
+		saveManager = gameController.GetComponent<SaveLoadGame> ();
 	}
 
 	public void NewGame(){
